feat(shared,document-model): move common types to shared
