+++ packages/document-drive/package.json
@@ -73,7 +73,7 @@
     "@powerhousedao/config": "workspace:*",
     "@powerhousedao/analytics-engine-core": "catalog:",
     "@prisma/client": "5.17.0",
-    "@sindresorhus/fnv1a": "^3.1.0",
+    "@sindresorhus/fnv1a": "catalog:",
     "change-case": "^5.4.4",
     "document-model": "workspace:*",
     "exponential-backoff": "^3.1.1",
