+++ packages/document-model/src/core/state.ts
@@ -1,5 +1,8 @@
+import {
+  HASH_ALGORITHM_SHA1,
+  HASH_ENCODING_BASE64,
+} from "shared/document-model";
 import type { PHAuthState, PHBaseState, PHDocumentState } from "./ph-types.js";
-import { HASH_ALGORITHM_SHA1, HASH_ENCODING_BASE64 } from "./ph-types.js";
 
 /**
  * Creates a default PHAuthState
