+++ packages/document-model/tsconfig.json
@@ -5,5 +5,10 @@
     "types": ["node", "vitest/globals", "jest"],
     "lib": ["ESNext", "DOM", "DOM.Iterable"],
   },
-  "include": ["**/*"]
+  "include": ["**/*"],
+  "references": [
+    {
+      "path": "../../shared"
+    }
+  ]
 }
