+++ packages/reactor/tsconfig.json
@@ -17,6 +17,9 @@
     },
     {
       "path": "../document-model"
+    },
+    {
+      "path": "../../shared"
     }
   ]
 }
