+++ shared/document-model/core/constants.ts
@@ -1,0 +1,8 @@
+// Known hash algorithms (can be extended without breaking changes)
+export const HASH_ALGORITHM_SHA1 = "sha1";
+export const HASH_ALGORITHM_SHA256 = "sha256";
+export const HASH_ALGORITHM_SHA512 = "sha512";
+
+// Known encodings (can be extended without breaking changes)
+export const HASH_ENCODING_BASE64 = "base64";
+export const HASH_ENCODING_HEX = "hex";
