+++ shared/document-model/core/ph-types.ts
@@ -5,4 +5,9 @@
 // Please do not add types to this file that have not been spec'd out in the
 // docs/planning folder in the reactor/ package.
 ///////////////////////////////////////////////////////////////////////////////
-export type * from "shared/document-model/core/ph-types";
+export type * from "./actions.ts";
+export type * from "./documents.ts";
+export type * from "./operations.ts";
+export type * from "./signatures.ts";
+export type * from "./state.ts";
+export type * from "./upgrades.ts";
