+++ shared/document-model/index.ts
@@ -1,0 +1,1 @@
+export * from "./core/constants.js";
