+++ shared/package.json
@@ -10,6 +10,22 @@
     ".": {
       "types": "./dist/index.d.ts",
       "default": "./dist/index.js"
+    },
+    "./document-model": {
+      "types": "./dist/document-model/index.d.ts",
+      "default": "./dist/document-model/index.js"
+    },
+    "./document-model/*": {
+      "types": "./dist/document-model/*.d.ts",
+      "default": "./dist/document-model/*.js"
     }
+  },
+  "devDependencies": {
+    "kysely": "catalog:",
+    "@sindresorhus/fnv1a": "catalog:",
+    "@powerhousedao/analytics-engine-core": "catalog:",
+    "mutative": "catalog:",
+    "react": "catalog:",
+    "@types/react": "catalog:"
   }
 }
